Updates to file functions, tokens, caching authtypes
