+++ lib/arrayFuncs.go
