+++ lib/fileFuncs.go
